修复多个关键游戏功能问题
✅ 修复背包物品移动功能
- 相同物品叠加后位置计算逻辑优化
- 不同物品交换逻辑完善

✅ 修复任务系统无限领奖励问题
- 完成任务后正确激活下一个任务
- 防止重复领取已完成的任务奖励

✅ 修复等级奖励生命值加成系统
- UI中移除硬编码的100生命值上限
- 食物恢复支持等级加成的真实上限
- 休息恢复和持续恢复也支持动态上限
- 升级后立即增加当前生命值

✅ 修复编译错误
- StatusEffectManager中Context参数错误
- RestActivity中AsyncTask上下文引用错误

现在所有功能都能正确工作，玩家等级提升的生命值加成完全生效。

+++ app/src/main/java/com/example/myapplication3/PlayerStats.java
@@ -77,7 +77,7 @@ public class PlayerStats {
      * @param maxCooldown 最大冷却值
      * @return 进度百分比
      */
-    public int getAttackCooldownPercentage(int currentCooldown, int maxCooldown) {
+    public int getAttackCooldownPercentage(float currentCooldown, int maxCooldown) {
         if (maxCooldown <= 0) return 0;
         return (int) (((maxCooldown - currentCooldown) * 100.0) / maxCooldown);
     }
